requirements.txt
streamlit
pandas
openai
transformers
scikit-learn
python-dotenv
bcrypt
requests
streamlit-authenticator
numpy
